+++ packages/block-editor/src/components/block-mobile-toolbar/index.native.js
@@ -23,17 +23,20 @@ const BlockMobileToolbar = ( {
 	clientId,
 	onDelete,
 	order,
-	horizontalDirection,
+	isStackedHorizontally,
 } ) => (
 	<View style={ styles.toolbar }>
 		<BlockMover
 			clientIds={ [ clientId ] }
-			horizontalDirection={ horizontalDirection }
+			isStackedHorizontally={ isStackedHorizontally }
 		/>
 
 		<View style={ styles.spacer } />
 
-		<BlockSettingsButton.Slot />
+		<BlockSettingsButton.Slot>
+			{ /* Render only one settings icon even if we have more than one fill - need for hooks with controls */ }
+			{ ( fills = [ null ] ) => fills[ 0 ] }
+		</BlockSettingsButton.Slot>
 
 		<ToolbarButton
 			title={ sprintf(