+++ packages/block-editor/src/components/block-mover/index.native.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-import { I18nManager } from 'react-native';
 import { first, last, partial, castArray } from 'lodash';
 
 /**
@@ -52,7 +51,7 @@ const BlockMover = ( {
 	onMoveUp,
 	firstIndex,
 	rootClientId,
-	horizontalDirection,
+	isStackedHorizontally,
 } ) => {
 	const {
 		backwardButtonIcon,
@@ -61,7 +60,7 @@ const BlockMover = ( {
 		forwardButtonHint,
 		firstBlockTitle,
 		lastBlockTitle,
-	} = horizontalDirection ? horizontalMover : verticalMover;
+	} = isStackedHorizontally ? horizontalMover : verticalMover;
 
 	if ( isLocked || ( isFirst && isLast && ! rootClientId ) ) {
 		return null;
@@ -72,7 +71,7 @@ const BlockMover = ( {
 		forwardButtonProp,
 		backwardButtonProp
 	) => {
-		if ( isRTL && horizontalDirection ) {
+		if ( isRTL && isStackedHorizontally ) {
 			// for RTL and horizontal direction switch prop between forward and backward button
 			if ( isBackwardButton ) {
 				return forwardButtonProp; // set forwardButtonProp for backward button
@@ -89,9 +88,10 @@ const BlockMover = ( {
 		const direction = isBackwardButton ? -1 : 1;
 		const toIndex = fromIndex + direction; // position after move
 
-		const { backwardButtonTitle, forwardButtonTitle } = horizontalDirection
-			? horizontalMover
-			: verticalMover;
+		const {
+			backwardButtonTitle,
+			forwardButtonTitle,
+		} = isStackedHorizontally ? horizontalMover : verticalMover;
 
 		const buttonTitle = switchButtonPropIfRTL(
 			isBackwardButton,
@@ -141,6 +141,7 @@ export default compose(
 			getTemplateLock,
 			getBlockRootClientId,
 			getBlockOrder,
+			getSettings,
 		} = select( 'core/block-editor' );
 		const normalizedClientIds = castArray( clientIds );
 		const firstClientId = first( normalizedClientIds );
@@ -156,7 +157,7 @@ export default compose(
 			firstIndex,
 			isFirst: firstIndex === 0,
 			isLast: lastIndex === blockOrder.length - 1,
-			isRTL: I18nManager.isRTL,
+			isRTL: getSettings().isRTL,
 			isLocked: getTemplateLock( rootClientId ) === 'all',
 			rootClientId,
 		};