+++ packages/editor/src/components/post-publish-button/index.js
@@ -16,7 +16,6 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import EntitiesSavedStates from '../entities-saved-states';
 import PublishButtonLabel from './label';
 
 export class PostPublishButton extends Component {
@@ -49,6 +48,13 @@ export class PostPublishButton extends Component {
 				this.setState( {
 					entitiesSavedStatesCallback: () => callback( ...args ),
 				} );
+				// Open the save panel by setting its callback.
+				// To set a function on the useState hook, we must set it
+				// with another function (() => myFunction). Passing the
+				// function on its own will cause an error when called.
+				this.props.setEntitiesSavedStatesCallback(
+					() => this.closeEntitiesSavedStates
+				);
 				return noop;
 			}
 
@@ -96,7 +102,6 @@ export class PostPublishButton extends Component {
 			visibility,
 			hasNonPostEntityChanges,
 		} = this.props;
-		const { entitiesSavedStatesCallback } = this.state;
 
 		const isButtonDisabled =
 			isSaving ||
@@ -170,10 +175,6 @@ export class PostPublishButton extends Component {
 		const componentChildren = isToggle ? toggleChildren : buttonChildren;
 		return (
 			<>
-				<EntitiesSavedStates
-					isOpen={ Boolean( entitiesSavedStatesCallback ) }
-					onRequestClose={ this.closeEntitiesSavedStates }
-				/>
 				<Button
 					ref={ this.buttonNode }
 					{ ...componentProps }