+++ packages/create-block/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Master
 
+## 0.12.0 (2020-04-30)
+
+### New Features
+
+- Add more CLI options: `--namespace`, `--title`, `--short-description` and `--category`. The goal is to make it easier to override default values used for scaffolding ([#21751](https://github.com/WordPress/gutenberg/pull/21751)).
+
+### Enhancements
+
+- Update `esnext` template to scaffold 3 JavaScript source files to illustrate how ES modules help to better organize code.
+
+## 0.10.0 (2020-04-01)
+
 ### New Features
 
 - Added readme.txt file to the existing templates to make your entry in the plugin browser most useful ([#20694](https://github.com/WordPress/gutenberg/pull/20694)).